feat(formats): Implement Atom XML support
Requires updating from `xml.etree` in Python's standard library to the `lxml` extension
package, not for performance and this point but for convenience features it offers
regarding XML namespaces.  We want to be as tolerant of poorly behaved feeds as we can
but we also want to preserve the feed content as much as possible so we need to be
namespace agnostic and `lxml` provides that.