fix(url): Various remote vs archive URL fixes
When served as intended by a static site server such as nginx, the `Content-Type` header
is set by matching the file's extension to its `mime.types` file.  So if the
`Content-Type` header from the archive is to match that of the remote then the extension
needs to be correct.

By changing or adding extensions to the archive download paths, as would happen when the
remote URL doesn't have the correct extension, we break the round-trip of URL to archive
path and back again.  This largely doesn't matter, at least with currently known use
cases, but is worth noting.

Also includes keeping the `<link rel="self" href="..." ...` URL absolute so it can be
used as a proper BASE URL in clients.  Also includes various other URL-related fixes.

These changes got away from me a bit and that in and of itself probably means some
refactoring is needed.  The changes became too intertwined to pull apart into separate,
atomic commits without taking a ton more time than making the changes.  So I'm making a
pragmatic choice to commit this monster, move on, and start refactoring later or the
next time this happens.