feat(config): Switch to YAML for feed configs
We need a base URL to assemble absolute URLs where relative URLs can't be used, such as
the ``<link rel="self"...`` in the archived version of the feed so that clients can
correctly interpret relative URLs in the archived feed.  Trying to shoehorn that into
the CSV format is exactly the case I was waiting for before finally moving ahead with a
switch.  Allowing users to manage the archived feeds in spreadsheet software isn't known
to be valuable or important enough to be worth the clumsiness of supporting global/base
defaults in CSV.