fix(feed): Resume after interruption, write items
Write the feed XML to disk after each feed item is successfully handled.  This may be
somewhat expensive to do for each item but probably not and is definitely premature
optimization to prioritize that over the known and much more likely case where
processing a large number of feed items is interrupted.