fix(requests): Memory leak, was exhausting all
Aparently the default `CacheControl` implementation caches everything forever which
quickly exhausts memory when downloading large media files such as podcast episodes.  It
also ocurred to me that it makes no sense for this project to use it anyways because we
already write everthing we download into the archive and check for its existence before
downloading anyways.

This reverts commit c2827c1defe5c28482105133386ec233e64fa956.