fix(link): Filter out duplicate paths from plugins
Allow link path plugins to return duplicate paths and filter them out to avoid writing
redundant `content-link-#` attributes to the feed XML in the archive.  This happens, for
example, when the `sonarr` plugin handles a feed item enclosure that maps to multiple
episodes when the corresponding Series episodes are also in one video file.  One could
argue, then, that the `sonarr` plugin should not return those duplicates, but I could
imagine that de-duplication might be a helpful service to provide to plugins and
can't imagine any meaningful or valid reason a plugin might want to return duplicates.