feat(link)!: Switch content term to enclosure
Having thought about playlist generation and whether to use the existing `link-path`
plugin system, I came to realize that `content` is wrong term to use for item
enclosures.  From the perspective of podcasts the enclosures may be the primary content
of a feed, but from a more general perspective feed content may be any number of things,
including directly in the feed XML.  The term `enclosure`, OTOH, is meant to describe a
more specific case of content that it too large or otherwise inappropriate to include in
the feed XML and thus must be stored externally, which is exactly what the `link-path`
plugin system is meant to deal with.